flask/requirements.txt
flask>=2.0.0
flask-login>=0.6.0
requests>=2.28.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
python-dotenv>=1.0.0
